[MNT] Fix certain workflow actions for forks and add button for resetting CI (#1764)
* [MNT] Bot to make empty commits (#15)

* CI fun

* comments and token

* [MNT] empty commit PR 2 (#16)

* CI fun

* comments and token

* comments

* skip dirty check

* Empty commit for CI

* empty comment 3 (#17)

* comment

* prs only

* comment

* Empty commit for CI

* empty commit 4 (#18)

* comment

* correct branch

* Empty commit for CI

* empty commit 5 (#19)

* comment

* branch

* empty commit 6 (#20)

* [BUG] fix padder fit_values (#1752)

* fix padder

* declare var

* [DOC,MNT] Documentation updates and deprecation warning for `v1.0.0` (#1743)

* forecasting docs

* v1.0.0 mass deprecation message

* index fix

* installation

* typos

* comment

* pushing to forks is crazy

* dont create branches

* [MNT] Unit testing revamp part 1: `check_estimator` (#1479)

* remove no soft deps job and expand full no soft deps job

* no PR testing flag

* module soft deps

* is this actually returning an empty list

* import modules now

* clear out module soft dep

* check_estimator

* check_estimator updates

* import

* more complete functions and tests

* fix

* coverage and paramterize function test

* testing data start

* testing data start

* testing data cont

* real partial

* cont

* testing data changes

* examples revert

* fourier features

* Revert "fourier features"

This reverts commit 567c5e1a37d46ab7710a81e96d270c78ec579423.

* legacy skip

* init

* move to legacy

* example data update

* testing data

* imports

* examples

* testing update

* multiindex

* multiindex fixes

* multiindex returns

* test

* test

* docs, tests and fixes

* comments

* last bits

* refactor testing folder

* imports

* series generators

* fixes

* make_series legacy

* merge

* forecasting functions

* fixes

* fixes

* fixes

* testing

* examples

* convert checks

* fixes

* docs

* ad datasets

* more docs

* forecasting compose

* forecasting

* delete forecasting api docs

* docs

* revert and comment init

* imports

* import test

* uncomment

* docstrings

* no imports

* package name

* import

* docstring

* fixes

* remove legacy and doc fix

* [ENH] Channel selection rocket wrapper (#1504)

* base channel selector

* base channel selector

* random channel selector

* elbow class test

* random channel selector

* base channel selector

* base channel selector

* revert registry

* draft channel selector

* class method

* test channel scorer

* docstring

* remove unnecessary ExponentTransform import

* deprecate DateTime

* Revert "deprecate DateTime"

This reverts commit 98b29e68494f191a15e07fa0d24632d9398000e4.

* channel selection

* remove dependency

* move tests to fit

* move tests to fit

* channel scorer test

* add to API

* maint

* maint

---------


* empty commit 7 (#21)

* comment

* remove branch param

* fixes

* Empty commit for CI

* empty commit 8 (#22)

* comment

* repo and branch

* comment

* empty commit 9 (#23)

* [BUG] fix padder fit_values (#1752)

* fix padder

* declare var

* [DOC,MNT] Documentation updates and deprecation warning for `v1.0.0` (#1743)

* forecasting docs

* v1.0.0 mass deprecation message

* index fix

* installation

* typos

* [MNT] Unit testing revamp part 1: `check_estimator` (#1479)

* remove no soft deps job and expand full no soft deps job

* no PR testing flag

* module soft deps

* is this actually returning an empty list

* import modules now

* clear out module soft dep

* check_estimator

* check_estimator updates

* import

* more complete functions and tests

* fix

* coverage and paramterize function test

* testing data start

* testing data start

* testing data cont

* real partial

* cont

* testing data changes

* examples revert

* fourier features

* Revert "fourier features"

This reverts commit 567c5e1a37d46ab7710a81e96d270c78ec579423.

* legacy skip

* init

* move to legacy

* example data update

* testing data

* imports

* examples

* testing update

* multiindex

* multiindex fixes

* multiindex returns

* test

* test

* docs, tests and fixes

* comments

* last bits

* refactor testing folder

* imports

* series generators

* fixes

* make_series legacy

* merge

* forecasting functions

* fixes

* fixes

* fixes

* testing

* examples

* convert checks

* fixes

* docs

* ad datasets

* more docs

* forecasting compose

* forecasting

* delete forecasting api docs

* docs

* revert and comment init

* imports

* import test

* uncomment

* docstrings

* no imports

* package name

* import

* docstring

* fixes

* remove legacy and doc fix

* [ENH] Channel selection rocket wrapper (#1504)

* base channel selector

* base channel selector

* random channel selector

* elbow class test

* random channel selector

* base channel selector

* base channel selector

* revert registry

* draft channel selector

* class method

* test channel scorer

* docstring

* remove unnecessary ExponentTransform import

* deprecate DateTime

* Revert "deprecate DateTime"

This reverts commit 98b29e68494f191a15e07fa0d24632d9398000e4.

* channel selection

* remove dependency

* move tests to fit

* move tests to fit

* channel scorer test

* add to API

* [DEP] Exclusion list  (#1730)

* remove dependency

* excluded list

* [MNT] Removing `tensorflow_addons` dependency (#1421)

* tensorflow bound

* add manually instance normalization

* empty commit

* fix doxs

* fix bug

* add maintainer

* re arrange

* dep on utils

* fix comments

* add version and file to header

* re add conflict

* remove addon

* only adodn remove

* add typeguard dep

* remove addons from test

* remove tag of python<3.12 from base class

* edit pyptoject for test details

* remove typeguard

* remove addons

* remove typeguard

* fix bug in python version

* remove config on python < 3.12

* remove python version from test

* remove python version from base

* use group norm

* re-add 3.12 limit

* remove tags and add python version

* skip random state clr test

* fix test all networks

* set y for channel selection test

* set y for channel selection test

---------


* comment

* Empty commit for CI

* uncomment

---------


---------
