+++ .all-contributorsrc
@@ -2349,7 +2349,7 @@
         "code",
         "doc",
         "research",
-        "tests",
+        "test",
         "review",
         "data"
       ]
