+++ .github/workflows/periodic_github_maintenace.yml
@@ -23,7 +23,7 @@ jobs:
           private-key: ${{ secrets.PR_APP_KEY }}
 
       - name: Stale Branches
-        uses: crs-k/stale-branches@v5
+        uses: crs-k/stale-branches@v5.0.1
         with:
           repo-token: ${{ steps.app-token.outputs.token }}
           days-before-stale: 140
