+++ .github/workflows/precommit_autoupdate.yml
@@ -17,7 +17,7 @@ jobs:
         with:
           python-version: "3.10"
 
-      - uses: browniebroke/pre-commit-autoupdate-action@v1
+      - uses: browniebroke/pre-commit-autoupdate-action@v1.0.0
 
       - if: always()
         name: Create app token
